Summary of proposed updates and changes for User Guide
The official English Linux Mint user guide is out of date.

This pdf is a summary of additions I'd like to see added to the official Linux Mint user guide.

I have also edited the .xml files for the chapters I wanted to see updated. And I added a chapter for frequently asked questions. Finally, I added some screenshots to demonstrate concepts in the user guide. The screenshots are of my personal desktop, and may not be suitable for an official user guide, but I give permission to user them.